+++ outboundSender.go
@@ -142,7 +142,6 @@ type CaduceusOutboundSender struct {
 	droppedInvalidConfig             metrics.Counter
 	droppedPanic                     metrics.Counter
 	cutOffCounter                    metrics.Counter
-	contentTypeCounter               metrics.Counter
 	queueDepthGauge                  metrics.Gauge
 	renewalTimeGauge                 metrics.Gauge
 	deliverUntilGauge                metrics.Gauge
@@ -547,8 +546,6 @@ func (obs *CaduceusOutboundSender) send(urls *ring.Ring, secret, acceptType stri
 	body := payload
 	var payloadReader *bytes.Reader
 
-	obs.contentTypeCounter.With("content_type", strings.TrimLeft(msg.ContentType, "application/")).Add(1.0)
-
 	// Use the internal content type unless the accept type is wrp
 	contentType := msg.ContentType
 	switch acceptType {
@@ -703,8 +700,6 @@ func (obs *CaduceusOutboundSender) queueOverflow() {
 		req.Header.Set("X-Webpa-Signature", sig)
 	}
 
-	//  record content type, json.
-	obs.contentTypeCounter.With("content_type", "json").Add(1.0)
 	resp, err := obs.sender(req)
 	if nil != err {
 		// Failure