+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Made outgoing hostname validation configurable [#217](https://github.com/xmidt-org/caduceus/pull/217)
 
 ## [v0.2.8]
 ### Changed